chore(home): rebase and keep the existing homeview develop and renamed to home2view
